requirements.txt
linkedin-api>=2.2.0
python-dotenv>=1.2.0
